move to_triples_string from re to llm
